fix @louiesinadjan: complete:

+++ single-threaded-escg/visualise.hpp
@@ -1,12 +1,14 @@
 #pragma once
 #include "external/matplotlib-cpp/matplotlibcpp.h"
 #include <vector>
+#include <cmath>
+#include <filesystem> 
 #include "simulation.hpp" // Include for Species and grid data
 
 namespace plt = matplotlibcpp;
 
 // Function to plot densities over Monte Carlo steps
-void plot_densities(const std::vector<int>& steps,
+void plot_densities(const std::vector<double>& steps,
                     const std::vector<double>& densityRock,
                     const std::vector<double>& densityPaper,
                     const std::vector<double>& densityScissors,
